+++ include/hive/modelsLibrary/discoveredEntitiesModel.hpp
@@ -88,6 +88,7 @@ public:
 		// Static information
 		la::avdecc::UniqueIdentifier entityID{};
 		bool isAemSupported{ false };
+		bool hasAnyConfigurationTree{ false };
 		la::avdecc::UniqueIdentifier entityModelID{};
 		std::optional<QString> firmwareVersion{};
 		std::optional<la::avdecc::entity::model::MemoryObjectIndex> firmwareUploadMemoryIndex{ std::nullopt };